rename pages and toutes
