+++ packages/client/src/pages/error/404/index.ts
