+++ packages/client/src/pages/error/500/index.ts
