+++ packages/client/src/pages/error/500/ui/index.tsx
