+++ packages/client/src/shared/components/ui/BaseLink/BaseLink.spec.tsx
@@ -2,6 +2,7 @@ import { render, screen } from '@testing-library/react'
 
 import '@testing-library/jest-dom'
 import BaseLink from '@/shared/components/ui/BaseLink'
+import { paths } from '@/shared/config/routing'
 
 describe('<BaseLink />', () => {
   it('Snapshot is correct', () => {
@@ -10,7 +11,7 @@ describe('<BaseLink />', () => {
   })
 
   it('Renders correctly with default props', () => {
-    render(<BaseLink to="/" />)
+    render(<BaseLink to={paths.home} />)
     const button = screen.getByRole('a')
 
     expect(button).toMatchSnapshot()
